Refactor header munging
