+++ src/commands/dev/mod.rs
@@ -1,11 +1,13 @@
 mod server_config;
 use server_config::ServerConfig;
+mod headers;
+use headers::{prepend_request_headers_prefix, strip_response_headers_prefix};
 
 use chrono::prelude::*;
 
 use hyper::client::{HttpConnector, ResponseFuture};
-use hyper::header::{HeaderMap, HeaderName, HeaderValue};
-use hyper::http::response::Parts;
+use hyper::header::{HeaderName, HeaderValue};
+
 use hyper::http::uri::InvalidUri;
 use hyper::service::{make_service_fn, service_fn};
 use hyper::{Body, Client, Request, Response, Server, Uri};
@@ -23,7 +25,6 @@ use crate::settings::toml::Target;
 use crate::terminal::emoji;
 
 const PREVIEW_HOST: &str = "rawhttp.cloudflareworkers.com";
-const HEADER_PREFIX: &str = "cf-ew-raw-";
 
 pub async fn dev(
     target: Target,
@@ -57,7 +58,7 @@ pub async fn dev(
 
                     let (mut parts, body) = resp.into_parts();
 
-                    munge_response_headers(&mut parts)?;
+                    strip_response_headers_prefix(&mut parts)?;
 
                     let resp = Response::from_parts(parts, body);
                     Ok::<_, failure::Error>(resp)
@@ -74,27 +75,11 @@ pub async fn dev(
     Ok(())
 }
 
-fn munge_response_headers(parts: &mut Parts) -> Result<(), failure::Error> {
-    let mut headers = HeaderMap::new();
-
-    for header in &parts.headers {
-        let (name, value) = header;
-        let name = name.as_str();
-        if name.starts_with(HEADER_PREFIX) {
-            let header_name = &name[HEADER_PREFIX.len()..];
-            let header_name = HeaderName::from_bytes(header_name.as_bytes())?;
-            headers.insert(header_name, value.clone());
-        }
-    }
-    parts.headers = headers;
-    Ok(())
-}
-
 fn get_preview_url(path_string: &str) -> Result<Uri, InvalidUri> {
     format!("https://{}{}", PREVIEW_HOST, path_string).parse()
 }
 
-fn get_path_as_str(uri: Uri) -> String {
+fn get_path_as_str(uri: &Uri) -> String {
     uri.path_and_query()
         .map(|x| x.as_str())
         .unwrap_or("")
@@ -109,21 +94,12 @@ fn preview_request(
 ) -> ResponseFuture {
     let (mut parts, body) = req.into_parts();
 
-    let path = get_path_as_str(parts.uri);
+    let path = get_path_as_str(&parts.uri);
     let method = parts.method.to_string();
     let now: DateTime<Local> = Local::now();
     let preview_id = &preview_id;
 
-    let mut headers: HeaderMap = HeaderMap::new();
-
-    for header in &parts.headers {
-        let (name, value) = header;
-        let forward_header = format!("{}{}", HEADER_PREFIX, name);
-        let header_name = HeaderName::from_bytes(forward_header.as_bytes())
-            .expect(&format!("Could not create header name for {}", name));
-        headers.insert(header_name, value.clone());
-    }
-    parts.headers = headers;
+    prepend_request_headers_prefix(&mut parts);
 
     parts.headers.insert(
         HeaderName::from_static("host"),
